+++ src/parser/trees/visitors/CMakeLists.txt
@@ -5,5 +5,6 @@ add_library(meta-tree-visitors annotation_remover.cpp
                                debinarizer.cpp
                                empty_remover.cpp
                                head_finder.cpp
+                               leaf_node_finder.cpp
                                transition_finder.cpp
                                unary_chain_remover.cpp)
